+++ icegraph/trainer/callbacks/__init__.py
@@ -1,0 +1,10 @@
+# Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
+# Developed by Taylor St Jean
+
+from .models import CheckpointCallback, ConsoleCallback, TensorBoardCallback
+
+CheckpointCallback.__module__ = __name__
+ConsoleCallback.__module__ = __name__
+TensorBoardCallback.__module__ = __name__
+
+__all__ = ["CheckpointCallback", "ConsoleCallback", "TensorBoardCallback"]