+++ icegraph/trainer/callbacks/base/__init__.py
@@ -1,0 +1,8 @@
+# Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
+# Developed by Taylor St Jean
+
+from .models import Callback
+
+Callback.__module__ = __name__
+
+__all__ = ["Callback"]