+++ main.py
@@ -1,4 +1,5 @@
 import warnings
+from backcall import callback_prototype
 
 warnings.filterwarnings(
     "ignore",
@@ -8,37 +9,35 @@ warnings.filterwarnings(
 
 from pathlib import Path
 
-import pandas as pd
-
 from icegraph.data.processor import FeatureProcessor
 from icegraph.data.extractor import FeatureExtractor
 from icegraph.data import DatasetRegistry
 from icegraph.config import IGConfig
 from icegraph.data.splitter import DatasetSplitter
 from icegraph.trainer import Trainer
+from icegraph.trainer.callbacks import CheckpointCallback, ConsoleCallback, TensorBoardCallback
 
 
 def main():
-    config_path = "./config/config.yaml"
+    config_path = Path("./config/config.yaml")
     config = IGConfig(config_path)
 
     # register the config instance
     IGConfig.register(config)
 
     # specify the data input directory, usually the i3 file set
-    resource = "/data/i3store/users/tstjean/i3_100_test"
+    resource = [Path(f"/data/i3store/users/tstjean/i3_100_test/extraction/splits/{split}.graphs.lmdb") for split in ["train", "val", "test"]]
 
     # define the processing chain and run each process sequentially left to right
-    for stage in [FeatureExtractor, FeatureProcessor, DatasetSplitter]:
+    for stage in []:
         processor = stage(resource)
         resource = processor()
 
     # load all data
     dataset_registry = DatasetRegistry.load_from_lmdb(*resource)
 
-    # use dataset_registry to pass data to training system
     trainer = Trainer(dataset_registry)
-    trainer.run(True)
+    trainer.run()
 
 if __name__ == "__main__":
     main()